# face_recognition_api/app/api/v1/endpoints/face_recognition_routes.py
from __future__ import annotations

import json

from app.core.contants import ErrorCode
from app.core.contants import HTTPStatusCode
from app.core.db import SessionLocal
from app.db.user import create_user
from app.db.user import get_user_by_embedding
from app.services.detect_face_service import face_detector
from app.utils.image_handling import get_image
from app.utils.image_handling import open_image
from face_recognition.aura_face import create_embedding
from robyn import Request
from robyn import Response
from robyn import SubRouter
from robyn.types import FormData

router = SubRouter(__file__, prefix='/api/v1/face')


@router.post('/add')
async def add_face(request: Request, form_data: FormData):
    """
    Add a new face embedding to the database. Expects an image file, user_id,
    and group_id in the form data.
     - user_id: A unique identifier for the user.
     - group_id: An identifier for the group associated with the user.
     - image: An image file containing the face to be added.
    """

    image = get_image(request)
    if not image:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.INVALID_IMAGE.value})
        )

    user_id: str = form_data.get('user_id', [None])
    group_id: str = form_data.get('group_id', [None])

    print(f"user_id: {user_id}",
          f"group_id: {group_id}")
    if not user_id or not group_id:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps(
                {'error': ErrorCode.MISSING_USER_OR_GROUP_ID.value})
        )
    image = open_image(image)
    faces = face_detector.find_faces(image)

    if not faces:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.NO_FACE.value})
        )

    if len(faces) > 1:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.MULTIPLE_FACES.value})
        )

    face = faces[0]
    left, top, right, bottom = face['bbox']
    face_image = image.crop((left, top, right, bottom))

    embeddings = create_embedding(face_image)
    with SessionLocal() as db:
        new_user = create_user(db, user_id, group_id, embeddings)

    if new_user is None:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.INTERNAL_SERVER_ERROR.value,
            description=json.dumps(
                {'error': ErrorCode.FAILED_TO_CREATE_USER.value})
        )


@router.post('/identify')
async def identify_face(request: Request):
    """
    Identify a face from the uploaded image. Expects an image file in the request.
     - image: An image file containing the face to be identified.
    Returns the user_id and group_id associated with the identified face,
    or an error if the face is not found or the image is invalid.
     - user_id: The unique identifier for the user associated with the identified face.
     - group_id: The identifier for the group associated with the identified face.
    """

    image = get_image(request)
    if not image:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.INVALID_IMAGE.value})
        )
    image = open_image(image)
    faces = face_detector.find_faces(image)
    if not faces:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.NO_FACE.value})
        )

    if len(faces) > 1:
        return Response(
            headers={'Content-Type': 'application/json'},
            status_code=HTTPStatusCode.BAD_REQUEST.value,
            description=json.dumps({'error': ErrorCode.MULTIPLE_FACES.value})
        )

    face = faces[0]
    left, top, right, bottom = face['bbox']
    face_image = image.crop((left, top, right, bottom))

    embeddings = create_embedding(face_image)
    with SessionLocal() as db:
        user = get_user_by_embedding(db, embeddings)
    if user is None:
        return {'message': ErrorCode.USER_NOT_FOUND.value}
    return {'user_id': user.user_id, 'group_id': user.group_id}