# face_recognition_api/openapi_local.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Face Recognition API",
    "version": "1.0.0",
    "description": "API for adding and identifying face embeddings."
  },

  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Robyn Development Server"
    }
  ],
  "paths": {
    "/api/v1/face/add": {
      "post": {
        "summary": "Add Face",
        "description": "Add a new face embedding to the database.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "A unique identifier for the user."
                  },
                  "group_id": {
                    "type": "string",
                    "description": "An identifier for the group associated with the user."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "An image file containing the face to be added."
                  }
                },
                "required": ["user_id", "group_id", "image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad Request (e.g., Invalid image or missing ID)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/face/identify": {
      "post": {
        "summary": "Identify Face",
        "description": "Identify a face from the uploaded image.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "An image file containing the face to be identified."
                  }
                },
                "required": ["image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful identification or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string"
                    },
                    "group_id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}